+++ src/domain-model/contracts/domain-rule-sequence.js
@@ -1,0 +1,9 @@
+export class DomainRuleSequence {
+  constructor({ steps = [], rulesEvaluatedBeforeStateChange = true, rulesEvaluatedBeforeDomainEvent = true }) {
+    this.steps = Object.freeze([...steps]);
+    this.rulesEvaluatedBeforeStateChange = Boolean(rulesEvaluatedBeforeStateChange);
+    this.rulesEvaluatedBeforeDomainEvent = Boolean(rulesEvaluatedBeforeDomainEvent);
+
+    Object.freeze(this);
+  }
+}